backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pyjwt==2.8.0
cryptography==41.0.7
bcrypt==4.0.1
httpx==0.27.0
psycopg2-binary==2.9.9